# DocsDev/agent/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LojaSync",
    "version": "1.2.3"
  },
  "paths": {
...
    "/products": {
      "get": {
        "summary": "List Products",
        "operationId": "list_products_products_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Product",
        "operationId": "create_product_products_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductPayload"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductItemResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Clear Products",
        "operationId": "clear_products_products_delete",
        "parameters": [
          {
            "name": "dry_run",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Dry Run"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Clear Products Products Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/brands": {
      "get": {
        "summary": "List Brands",
        "operationId": "list_brands_brands_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BrandsResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Add Brand",
        "operationId": "add_brand_brands_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BrandPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BrandsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BrandPayload": {
        "properties": {
          "nome": {
            "type": "string",
            "minLength": 1,
            "title": "Nome"
          }
        },
        "type": "object",
        "required": [
          "nome"
        ],
        "title": "BrandPayload"
      },
      "BrandsResponse": {
        "properties": {
          "marcas": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Marcas"
          }
        },
        "type": "object",
        "required": [
          "marcas"
        ],
        "title": "BrandsResponse"
      },
...
      "CorItemPayload": {
        "properties": {
          "cor": {
            "type": "string",
            "minLength": 1,
            "title": "Cor"
          },
          "quantidade": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Quantidade"
          }
        },
        "type": "object",
        "required": [
          "cor",
          "quantidade"
        ],
        "title": "CorItemPayload"
      },
...
      "GradeItemPayload": {
        "properties": {
          "tamanho": {
            "type": "string",
            "minLength": 1,
            "title": "Tamanho"
          },
          "quantidade": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Quantidade"
          }
        },
        "type": "object",
        "required": [
          "tamanho",
          "quantidade"
        ],
        "title": "GradeItemPayload"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProductItemResponse": {
        "properties": {
          "item": {
            "$ref": "#/components/schemas/ProductResponse"
          }
        },
        "type": "object",
        "required": [
          "item"
        ],
        "title": "ProductItemResponse"
      },
      "ProductListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ProductResponse"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "ProductListResponse"
      },
...
      "ProductPayload": {
        "properties": {
          "nome": {
            "type": "string",
            "minLength": 1,
            "title": "Nome"
          },
          "codigo": {
            "type": "string",
            "title": "Codigo",
            "default": ""
          },
          "quantidade": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Quantidade",
            "default": 1
          },
          "preco": {
            "type": "string",
            "minLength": 1,
            "title": "Preco"
          },
          "categoria": {
            "type": "string",
            "title": "Categoria",
            "default": ""
          },
          "marca": {
            "type": "string",
            "title": "Marca",
            "default": ""
          },
          "preco_final": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preco Final"
          },
          "descricao_completa": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Descricao Completa"
          },
          "grades": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/GradeItemPayload"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grades"
          },
          "cores": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/CorItemPayload"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cores"
          }
        },
        "type": "object",
        "required": [
          "nome",
          "preco"
        ],
        "title": "ProductPayload"
      },
      "ProductResponse": {
        "properties": {
          "nome": {
            "type": "string",
            "title": "Nome"
          },
          "codigo": {
            "type": "string",
            "title": "Codigo"
          },
          "codigo_original": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Codigo Original"
          },
          "quantidade": {
            "type": "integer",
            "title": "Quantidade"
          },
          "preco": {
            "type": "string",
            "title": "Preco"
          },
          "categoria": {
            "type": "string",
            "title": "Categoria"
          },
          "marca": {
            "type": "string",
            "title": "Marca"
          },
          "preco_final": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preco Final"
          },
          "descricao_completa": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Descricao Completa"
          },
          "grades": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/GradeItemPayload"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grades"
          },
          "cores": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/CorItemPayload"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cores"
          },
          "source_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Type"
          },
          "import_batch_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Import Batch Id"
          },
          "import_source_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Import Source Name"
          },
          "pending_grade_import": {
            "type": "boolean",
            "title": "Pending Grade Import",
            "default": false
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "ordering_key": {
            "type": "string",
            "title": "Ordering Key"
          }
        },
        "type": "object",
        "required": [
          "nome",
          "codigo",
          "codigo_original",
          "quantidade",
          "preco",
          "categoria",
          "marca",
          "preco_final",
          "descricao_completa",
          "timestamp",
          "ordering_key"
        ],
        "title": "ProductResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}